Fix warnings.
I know it's a minor thing, but it's kind of nice to have things compile without
warnings. This should make it easier to find other issues that show up. That
is, if you see a warning now, it's likely that it is new code that added it.

This corresponds to patch 3196483 (with two additions).

+++ src/dialogueloader_tlk.cpp
@@ -57,7 +57,7 @@ Dialogue* U4TlkDialogueLoader::load(void *source) {
     strings[2][0] = tolower(strings[2][0]);
 
     // ... then replace any newlines in the string with spaces
-    int index = strings[2].find ("\n");
+    size_t index = strings[2].find ("\n");
     while (index != string::npos)
     {
         strings[2][index] = ' ';
